Refactored to use functions instead of `attr`.
Depending on `attr` to create body and fixture
was not a good idea. Not only I had to take care
of all possibilities, but attributes for body and
fixtures are all different. For now, body will be
created at the first time setting a shape.
Multile shapes can be attached to the body through
chaining. Also fixed a bug with private variable
in module pattern.